@@ -3,12 +3,12 @@
   "version": "21.6.11",
   "private": true,
   "dependencies": {
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router": "^6.3.0",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "^5.0.0",
-    "web-vitals": "^0.2.4"
+    "web-vitals": "^3.0.0"
   },
   "scripts": {
     "start": "react-scripts start",
