@@ -11,18 +11,34 @@ const Header = (props) => {
       <div className="container">
         <div className="flex-header d-flex justify-content-between align-items-center">
           <div className="socials-list-hd s1 hv1">
-            <a href="https://www.facebook.com/shubham14p3" target="_blank" rel="noreferrer noopener">
+            <a
+              href="https://www.facebook.com/shubham14p3"
+              target="_blank"
+              rel="noreferrer noopener"
+            >
               <i className="fa fa-facebook" aria-hidden="true" />
             </a>
-            <a href="https://www.linkedin.com/in/shubham14p3/" target="_blank" rel="noreferrer noopener">
+            <a
+              href="https://www.linkedin.com/in/shubham14p3/"
+              target="_blank"
+              rel="noreferrer noopener"
+            >
               <i className="fa fa-linkedin" aria-hidden="true" />
             </a>
-            <a href="https://github.com/shubham14p3" target="_blank" rel="noreferrer noopener">
+            <a
+              href="https://github.com/shubham14p3"
+              target="_blank"
+              rel="noreferrer noopener"
+            >
               <i className="fa fa-github" aria-hidden="true" />
             </a>
-            <a href="https://wa.me/918092766575" target="_blank" rel="noreferrer noopener" >
-            <i className="fa fa-whatsapp" aria-hidden="true" />
-          </a>
+            <a
+              href="https://wa.me/918092766575"
+              target="_blank"
+              rel="noreferrer noopener"
+            >
+              <i className="fa fa-whatsapp" aria-hidden="true" />
+            </a>
           </div>
           <div className="content-menu d-lg-flex">
             <div className="nav-wrap">
@@ -37,7 +53,8 @@ const Header = (props) => {
                         <li key={data.id}>
                           <Link
                             to={data.tolink}
-                            onClick={() => {
+                            onClick={(e) => {
+                              e.preventDefault();
                               window.location.href = data.tolink;
                             }}
                             className={data.id === 1 ? "active" : ""}
@@ -57,27 +74,29 @@ const Header = (props) => {
                   ))}
                   {/* Disabled Blog from Main Header Navbar */}
                   {/* <li>
-                    <Link to="#">Blog</Link>
-                    <ul className="sub-menu">
-                      {bloglinks.map((data) => (
-                        <li key={data.id}>
-                          <Link
-                            to={data.toblog}
-                            onClick={() => {
-                              window.location.href = data.toblog;
-                            }}
-                          >
-                            {data.nameblog}
-                          </Link>
-                        </li>
-                      ))}
-                    </ul>
-                  </li> */}
+                  <Link to="#">Blog</Link>
+                  <ul className="sub-menu">
+                    {bloglinks.map((data) => (
+                      <li key={data.id}>
+                        <Link
+                          to={data.toblog}
+                          onClick={() => {
+                            window.location.href = data.toblog;
+                          }}
+                        >
+                          {data.nameblog}
+                        </Link>
+                      </li>
+                    ))}
+                  </ul>
+                </li> */}
                 </ul>
               </nav>
             </div>
             <div className="hire-me-s1 btn-general btn-hv-common d-lg-flex align-items-center">
-              <a href="/resume/resume.pdf" className="btn-inner border-corner2 lt-sp08 text-white"
+              <a
+                href="/resume/resume.pdf"
+                className="btn-inner border-corner2 lt-sp08 text-white"
               >
                 Resume
               </a>
